Move the executor abstraction layer to the actor level
The next step will be to introduce alternative executor using the
active-job like layer, where it's still beneficial to reuse the
orchestrator actor part.
